Check that the binaries were installed
Any argument that is passed to the binaries gives a non-zero exit code.
Since we do want to know whether the binary was installed or not,
we check whether there is an error or not.

If no binary exists, it'll give us an error.
If the binary exists, it won't try to execute it.

Ideally, the binaries would have something like `--version`
that would give a zero exit code.